Switch to 32-bit floats ('pub type Float = f32') and fix some bugs.

+++ examples/api_pbrt_world_end.rs
@@ -29,7 +29,7 @@ fn main() {
                                 35.0,
                                 String::from("pbrt.exr"),
                                 1.0,
-                                std::f64::INFINITY);
+                                std::f32::INFINITY);
     // TODO: std::unique_ptr<Scene> scene(renderOptions->MakeScene());
     // MakeAccelerator()
     // CreateBVHAccelerator()
